moved get_labelled_atlas to nifti cli file
